Move helper methods into respective util files
